Merge to main after releasing new version 0.8.0

+++ MyPo.Blazor/MyPo.Blazor.Portfolio.App/Shared/PortfolioUtils.cs
@@ -182,7 +182,7 @@ public static partial class PortfolioUtils
 			return 0;
 		}
 		var currentPrice = quote.MarketPrice;
-		var pnl = ((decimal)currentPrice - asset.AveragePrice*market!.PriceScale) * asset.Quantity;
+		var pnl = (currentPrice - asset.AveragePrice*market!.PriceScale) * asset.Quantity;
 		return pnl;
 	}
